requirements.txt
wtftz==0.2.4
pymongo==3.11.3
discord==1.0.1
python-dotenv==0.15.0
